# api/src/services/swagger_output.json
{
  "openapi": "3.1.1",
  "info": {
    "version": "1.0.0",
    "title": "Bingo Player API",
    "description": "Documentation of the Bingo Player API"
  },
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
  "paths": {
...
    "/api/v1/spaces/get/by-id/{id}": {
      "get": {
        "tags": [
          "Spaces",
          "Get"
        ],
        "description": "Get the space with the exact given ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the space to get."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSpaceByExactID200ResponseBody"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/GetSpaceByExactID200ResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              }
            }
          },
          "404": {
            "description": "No matching spaces found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/spaces/count/by-prop/{prop}/{query}": {
      "get": {
        "tags": [
          "Spaces",
          "Count",
          "Query"
        ],
        "description": "Get the number of spaces in the database matching the given query.",
        "parameters": [
          {
            "name": "prop",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SpaceQueryPropNonId"
            },
            "description": "The space query prop to search.",
            "type": "string"
          },
          {
            "name": "query",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The query string to search for."
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/o"
          },
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/l"
          },
          {
            "$ref": "#/components/parameters/caseSensitive"
          },
          {
            "$ref": "#/components/parameters/cs"
          },
          {
            "$ref": "#/components/parameters/accentSensitive"
          },
          {
            "$ref": "#/components/parameters/as"
          },
          {
            "$ref": "#/components/parameters/matchWhole"
          },
          {
            "$ref": "#/components/parameters/mw"
          },
          {
            "$ref": "#/components/parameters/invert"
          },
          {
            "$ref": "#/components/parameters/i"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountSpacesByProp200ResponseBody"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/CountSpacesByProp200ResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              }
            }
          },
          "404": {
            "description": "No matching spaces found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/spaces/delete/by-id/{id}": {
      "delete": {
        "tags": [
          "Spaces",
          "Delete"
        ],
        "description": "Delete a space by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the space to delete."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSpaceByID200ResponseBody"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSpaceByID200ResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              }
            }
          },
          "404": {
            "description": "No matching spaces found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/spaces/duplicate/by-id/{id}": {
      "post": {
        "tags": [
          "Spaces",
          "Duplicate"
        ],
        "description": "Duplicate an existing space in the database.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the space to duplicate."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateSpaceByID200ResponseBody"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateSpaceByID200ResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse400"
                }
              }
            }
          },
          "404": {
            "description": "No matching spaces found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse404"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse501"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "limit": {
        "schema": {
          "$ref": "#/components/schemas/limit"
        },
        "required": false,
        "name": "limit",
        "in": "query"
      },
      "l": {
        "schema": {
          "$ref": "#/components/schemas/l"
        },
        "required": false,
        "name": "l",
        "in": "query"
      },
      "offset": {
        "schema": {
          "$ref": "#/components/schemas/offset"
        },
        "required": false,
        "name": "offset",
        "in": "query"
      },
      "o": {
        "schema": {
          "$ref": "#/components/schemas/o"
        },
        "required": false,
        "name": "o",
        "in": "query"
      },
      "caseSensitive": {
        "schema": {
          "$ref": "#/components/schemas/caseSensitive"
        },
        "required": false,
        "name": "caseSensitive",
        "in": "query"
      },
      "cs": {
        "schema": {
          "$ref": "#/components/schemas/cs"
        },
        "required": false,
        "name": "cs",
        "in": "query"
      },
      "accentSensitive": {
        "schema": {
          "$ref": "#/components/schemas/accentSensitive"
        },
        "required": false,
        "name": "accentSensitive",
        "in": "query"
      },
      "as": {
        "schema": {
          "$ref": "#/components/schemas/as"
        },
        "required": false,
        "name": "as",
        "in": "query"
      },
      "matchWhole": {
        "schema": {
          "$ref": "#/components/schemas/matchWhole"
        },
        "required": false,
        "name": "matchWhole",
        "in": "query"
      },
      "mw": {
        "schema": {
          "$ref": "#/components/schemas/mw"
        },
        "required": false,
        "name": "mw",
        "in": "query"
      },
...
      "invert": {
        "schema": {
          "$ref": "#/components/schemas/invert"
        },
        "required": false,
        "name": "invert",
        "in": "query"
      },
      "i": {
        "schema": {
          "$ref": "#/components/schemas/i"
        },
        "required": false,
        "name": "i",
        "in": "query"
      }
    },
    "schemas": {
...
      "CleanMongoSpace": {
        "type": "object",
        "properties": {
          "name": {
            "$ref": "#/components/schemas/SpaceName"
          },
          "description": {
            "$ref": "#/components/schemas/SpaceDescription"
          },
          "examples": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpaceExample"
            }
          },
          "aliases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpaceName"
            },
            "description": "Aliases for the space.",
            "example": [
              "Alias 1",
              "Alias 2"
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpaceTag"
            }
          },
          "_id": {
            "$ref": "#/components/schemas/SpaceID"
          }
        },
        "required": [
          "name",
          "description",
          "examples",
          "aliases",
          "tags",
          "_id"
        ],
        "description": "The MongoDB representation of a Space."
      },
...
      "SpaceQueryPropNonId": {
        "type": "string",
        "enum": [
          "name",
          "description",
          "examples",
          "aliases",
          "tags",
          "known-as",
          "alias",
          "tag",
          "example"
        ],
        "description": "A property of a space that can be queried, excluding id."
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "UNKNOWN",
          "BAD_INPUT",
          "BAD_URL",
          "BAD_QUERY",
          "BAD_BODY",
          "INTERNAL",
          "NOT_FOUND",
          "NOT_IMPLEMENTED",
          "VALIDATION"
        ],
        "description": "An error code that may be returned by the server."
      },
      "ErrorMessage": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          {
            "type": "null"
          }
        ],
        "description": "A message describing an error.",
        "example": "Internal server error."
      },
      "HelpLink": {
        "type": "string",
        "format": "uri",
        "description": "A URL to the docs page that may help the user resolve the error or understand the response.",
        "example": "https://example.com/docs"
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ],
            "description": "Whether the operation was successful.",
            "example": false
          },
          "help": {
            "$ref": "#/components/schemas/HelpLink"
          },
          "code": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "description": "The error code.",
                "example": "UNKNOWN"
              }
            ]
          },
          "message": {
            "$ref": "#/components/schemas/ErrorMessage"
          }
        },
        "required": [
          "ok",
          "code",
          "message"
        ]
      },
      "ErrorResponse501": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ],
            "description": "Whether the operation was successful.",
            "example": false
          },
          "help": {
            "$ref": "#/components/schemas/HelpLink"
          },
          "code": {
            "type": "string",
            "enum": [
              "NOT_IMPLEMENTED"
            ],
            "description": "The error code."
          },
          "message": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorMessage"
              },
              {
                "example": "This feature is not yet implemented."
              }
            ]
          }
        },
        "required": [
          "ok",
          "code",
          "message"
        ]
      },
      "ErrorResponse400": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ],
            "description": "Whether the operation was successful.",
            "example": false
          },
          "help": {
            "$ref": "#/components/schemas/HelpLink"
          },
          "code": {
            "type": "string",
            "enum": [
              "BAD_INPUT",
              "BAD_BODY",
              "BAD_QUERY",
              "BAD_URL"
            ],
            "description": "The error code."
          },
          "message": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorMessage"
              },
              {
                "example": "Invalid input."
              }
            ]
          }
        },
        "required": [
          "ok",
          "code",
          "message"
        ]
      },
      "ErrorResponse404": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              false
            ],
            "description": "Whether the operation was successful.",
            "example": false
          },
          "help": {
            "$ref": "#/components/schemas/HelpLink"
          },
          "code": {
            "type": "string",
            "enum": [
              "NOT_FOUND"
            ],
            "description": "The error code."
          },
          "message": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ErrorMessage"
              },
              {
                "example": "No resources found."
              }
            ]
          }
        },
        "required": [
          "ok",
          "code",
          "message"
        ]
      },
      "SuccessResponseBase": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "Whether the operation was successful.",
            "example": true
          },
          "help": {
            "$ref": "#/components/schemas/HelpLink"
          }
        },
        "required": [
          "ok"
        ],
        "description": "A successful response from the server."
      },
...
      "offset": {
        "type": [
          "integer",
          "null"
        ],
        "minimum": 0,
        "default": 0,
        "description": "[Query Parameter] The number of spaces to skip before returning results. Must be a non-negative integer. If not provided, it will be considered 0."
      },
      "o": {
        "type": [
          "integer",
          "null"
        ],
        "minimum": 0,
        "default": 0,
        "description": "[Query Parameter] Shorthand for offset."
      },
...
      "GetSpaceByExactID200ResponseBody": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "space": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CleanMongoSpace"
                  },
                  {
                    "description": "The space that matches the query."
                  }
                ]
              }
            },
            "required": [
              "space"
            ],
            "description": "The response from getting spaces."
          },
          {
            "$ref": "#/components/schemas/SuccessResponseBase"
          }
        ]
      },
...
      "CountSpacesByProp200ResponseBody": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "count": {
                "type": "number",
                "minimum": 0,
                "description": "The number of spaces that match the query."
              }
            },
            "required": [
              "count"
            ]
          },
          {
            "$ref": "#/components/schemas/SuccessResponseBase"
          }
        ],
        "description": "The response from counting spaces."
      },
...
      "DeleteSpaceByID200ResponseBody": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "deleted": {
                "type": "boolean",
                "enum": [
                  true
                ],
                "description": "Whether the space was successfully deleted."
              }
            },
            "required": [
              "deleted"
            ],
            "description": "The response from deleting a space."
          },
          {
            "$ref": "#/components/schemas/SuccessResponseBase"
          }
        ]
      },
...
      "DuplicateSpaceByID200ResponseBody": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "space": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CleanMongoSpace"
                  },
                  {
                    "description": "The duplicated space."
                  }
                ]
              }
            },
            "required": [
              "space"
            ],
            "description": "The response from duplicating a space."
          },
          {
            "$ref": "#/components/schemas/SuccessResponseBase"
          }
        ]
      },
...
    }
  }
}